remove DebuggingUtil moved logPoints to LoggerService
