+++ src/main/java/com/wecca/canoeanalysis/models/canoe/HullSection.java
@@ -10,8 +10,8 @@ import com.wecca.canoeanalysis.models.load.ContinuousLoadDistribution;
 import com.wecca.canoeanalysis.models.load.LoadType;
 import com.wecca.canoeanalysis.models.function.Section;
 import com.wecca.canoeanalysis.models.function.VertexFormParabolaFunction;
+import com.wecca.canoeanalysis.services.LoggerService;
 import com.wecca.canoeanalysis.utils.CalculusUtils;
-import com.wecca.canoeanalysis.utils.DebuggingUtils;
 import com.wecca.canoeanalysis.utils.SharkBaitHullLibrary;
 import com.wecca.canoeanalysis.utils.PhysicalConstants;
 import lombok.EqualsAndHashCode;
@@ -258,7 +258,7 @@ public class HullSection extends Section
      * Note that this means the topmost point of the hull on the y-axis is y = 0
      */
     private void validateSign(Function<Double, Double> profileCurve, boolean positive) {
-        DebuggingUtils.logPoints(profileCurve, new Section(x, rx));
+        LoggerService.logPoints(profileCurve, new Section(x, rx));
 
         // Convert the hullShapeFunction to BoundedUnivariateFunction for compatibility with Apache Commons Math
         // Need to negate the function as BrentOptimizer finds the min, and we want the max
